+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * Plots use mono-spaced font and fixed tick label axis padding, to avoid jumping y axis (#26)
 * Remove the last tick label from the x axis if close to the right margin, to avoid jumping x axis (#29)
 * Scatter plots use solid instead of empty circle (#30)
+* `Window` does not call `Drawer.Draw` when it is minimized (#32)
 
 ## [[v0.2.0]](https://github.com/mlange-42/arche-pixel/compare/v0.1.0...v0.2.0)